pyproject.toml
[tool.poetry]
name = "langchain-opengauss"
version = "0.1.2"
description = "openGauss vector store integration for LangChain"
authors = []
readme = "README.md"
repository = "https://github.com/mpb159753/langchain-opengauss"
license = "MIT"

[tool.poetry.urls]
"Source Code" = "https://github.com/mpb159753/langchain-opengauss"

[tool.poetry.dependencies]
python = "^3.9"
langchain-core = ">=0.2.13,<0.4.0"
langchain-community = ">=0.3.20"
psycopg2-binary = ">=2.9.6"
pydantic = ">=2.9.0"


[tool.coverage.run]
omit = ["tests/*"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.codespell]
skip = '.git,*.pdf,*.svg,*.pdf,*.yaml,*.ipynb,poetry.lock,*.min.js,*.css,package-lock.json,example_data,_dist,examples,templates,*.trig'
